Version 1.1.8
Improved gameplay, added in more animations, fixed some bugs.

+++ src/entities/Dialog.java
@@ -13,9 +13,9 @@ public class Dialog {
 	
 	public static ArrayList<String> text = null;
 	public static int dialog_pos = -1;
-	public static int NPC_ID = -1;
+	public static Long NPC_ID = -1l;
 	
-	public static void CreateDialog(World world, ArrayList<String> dialog, int id, int x, int y) {
+	public static void CreateDialog(World world, ArrayList<String> dialog, long id, int x, int y) {
 		if (dialog != null) {
 			text = new ArrayList<String>(dialog);
 			dialog_pos = 0;
@@ -35,7 +35,7 @@ public class Dialog {
 				for (int i = 0; i < world.entities.size(); i++) {
 					if (world.entities.get(i).type == EntityDictionary.DIALOG_BOX) {
 						world.RemoveEntity(i);
-						Dialog.NPC_ID = -1;
+						Dialog.NPC_ID = -1l;
 						Main.SetState(States.RUNNING);
 						return false;
 					}
